+++ src/api/process/get.ts
@@ -1,10 +1,16 @@
 import * as vscode from "vscode"
+import * as fs from "fs"
 import { log, notif } from "../../utils/utils"
-import { MODEL_DEFAULT, COMMENT_SYMBOLS, MODELS_DEFAULT, ARGS_SUPPORTED, CONVERT_DEFAULT } from "../../utils/defaults"
+import {
+	MODEL_DEFAULT,
+	COMMENT_SYMBOLS,
+	MODELS_DEFAULT,
+	ARGS_SUPPORTED,
+} from "../../utils/defaults"
 
 export function getPrompt(prompt: vscode.TextLine) {
 	if (prompt.isEmptyOrWhitespace || !prompt) {
-		notif("Your prompt seems to be empty...")
+		notif("AI-Keys: Your prompt seems to be empty...")
 		log("AI-Keys: No prompt")
 
 		vscode.commands.executeCommand("aikeys.sendBoxPrompt")
@@ -35,92 +41,28 @@ export const getComment = () => {
 
 export function getModel(promptArray: string[], fullList = [""]) {
 	let model = ""
-	
-	promptArray.some(word => {
+
+	promptArray.some((word) => {
 		if (fullList.includes(word)) model = word
 		if (Object.keys(MODELS_DEFAULT).includes(word)) model = MODELS_DEFAULT[word]
 	})
 
 	if (model.length > 0) return model
 
-	notif(`Unrecognised model, switching to default ${MODELS_DEFAULT[MODEL_DEFAULT]}`)
-	log(`Model not recognised`)
+	notif(`AI-Keys: Unrecognised model, switching to default ${MODELS_DEFAULT[MODEL_DEFAULT]}`)
+	log(`AI-Keys: Model not recognised`)
 
 	return MODELS_DEFAULT[MODEL_DEFAULT]
 }
 
-export function getArg(promptArray: string[], prompt: string) {
-	const language = getLanguage() as string
-	let arg = promptArray.map(word => {
-		if (Object.values(ARGS_SUPPORTED).some(argArray => 
-				argArray.includes(word))) return word
-	}).join("")
-
-	if (ARGS_SUPPORTED.op.some(phrase => { return arg === phrase })) {
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `refactor and optimize this ${language} code \n${prompt}`
-	}
-
-	if (ARGS_SUPPORTED.cv.some(phrase => { return arg === phrase })) {
-		let convertTo = CONVERT_DEFAULT;
-		Object.keys(COMMENT_SYMBOLS).some(lang => {
-			if (prompt.includes(lang)) {
-				convertTo = lang
-				prompt = prompt.replace(lang, "").trim()
-				return true
-			}
+export function getArg(promptArray: string[]) {
+	return promptArray
+		.filter((word) => {
+			if (Object.values(ARGS_SUPPORTED).some((argArray) => argArray.includes(word))) return word
 		})
-
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `convert this from ${language} to ${convertTo} \n${prompt}`
-	}
-
-	if (ARGS_SUPPORTED.ex.some(phrase => { return arg === phrase })) {
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `explain this ${language} code \n${prompt}`
-	}
-
-	return prompt.trim()
-
-	// return prompt
-	// 	.split(" ")
-	// 	// .filter(word => { return word !== arg })
-	// 	.join(" ")
+		.join("")
 }
 
-export function addArg(arg: string, prompt: string) {
-	const language = getLanguage() as string
-
-	if (!Object.values(ARGS_SUPPORTED).some((val) => { return val.some((phrase) => { return arg === phrase })}) ) {
-		notif("Invalid argument, using default settings.\nCheck AI-Keys list of suported arguments.")
-		log("AI-Keys: Invalid prompt argument")
-		
-		return prompt
-	}
-
-	if (ARGS_SUPPORTED.op.some(phrase => { return arg === phrase })) {
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `refactor and optimize this ${language} code \n${prompt}`
-	}
-
-	if (ARGS_SUPPORTED.cv.some(phrase => { return arg === phrase })) {
-		let convertTo = CONVERT_DEFAULT;
-		Object.keys(COMMENT_SYMBOLS).some(lang => {
-			if (prompt.includes(lang)) {
-				convertTo = lang
-				prompt = prompt.replace(lang, "").trim()
-				return true
-			}
-		})
-
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `convert this from ${language} to ${convertTo} \n${prompt}`
-	}
-
-	if (ARGS_SUPPORTED.ex.some(phrase => { return arg === phrase })) {
-		prompt = prompt.replace(arg, "").trim()
-		prompt = `explain this ${language} code \n${prompt}`
-	}
-
-	return prompt.trim()
+export function getLocalImage(prompt: string) {
+	return fs.readFileSync(prompt, "base64")
 }